Update Go static http benchmarks to use module system
