--- bench/static_http/hello_fast.go
@@ -1,18 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"github.com/valyala/fasthttp"
-)
-
-func main() {
-	h := requestHandler
-	if err := fasthttp.ListenAndServe("0.0.0.0:8080", h); err != nil {
-		log.Fatalf("Error in ListenAndServe: %s", err)
-	}
-}
-
-func requestHandler(ctx *fasthttp.RequestCtx) {
-	fmt.Fprintf(ctx, "Hello, world!")
-}
