+++ bench/static_http/hello_go/go.mod
@@ -1,0 +1,3 @@
+module github.com/DmitryOlshansky/hello_go
+
+go 1.22.2
