+++ bench/static_http/hello_go_fast/go.mod
@@ -1,0 +1,10 @@
+module github.com/DmitryOlshansky/hello_fast
+
+go 1.22.2
+
+require (
+	github.com/andybalholm/brotli v1.1.0 // indirect
+	github.com/klauspost/compress v1.17.6 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasthttp v1.52.0 // indirect
+)
